refactored the code after Olaf's review

+++ src/main/java/io/github/miwlodar/controller/NotesController.java
@@ -1,42 +1,34 @@
 //1 of 3 controllers - handling all CRUD operations for notes and returning pages accordingly
 package io.github.miwlodar.controller;
 
-import java.util.List;
+import io.github.miwlodar.entity.Note;
 import io.github.miwlodar.service.NotesService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import io.github.miwlodar.entity.Note;
+import org.springframework.web.bind.annotation.*;
+import java.util.List;
 
 @Controller
 @RequestMapping("/notes")
 public class NotesController {
 
 	private final NotesService notesService;
-	
+
 	public NotesController(NotesService notesService) {
 		this.notesService = notesService;
 	}
 
 	@GetMapping("/list")
 	public String listNotes(Model model) {
-
 		List<Note> notes = notesService.findAll();
-
 		model.addAttribute("notes", notes);
-		
+
 		return "/notes/list-notes";
 	}
 	
 	@GetMapping("/form-for-add")
 	public String formForAdd(Model model) {
-
 		Note note = new Note();
-		
 		model.addAttribute("notes", note);
 		
 		return "/notes/notes-form";
@@ -44,11 +36,9 @@ public class NotesController {
 
 
 	@GetMapping("/form-for-update")
-	public String formForUpdate(@RequestParam("noteId") int id,
-									Model model) {
-
+	public String formForUpdate(@RequestParam("noteId") Long id,
+								Model model) {
 		Note note = notesService.findById(id);
-
 		model.addAttribute("notes", note);
 
 		return "/notes/notes-form";
@@ -57,7 +47,6 @@ public class NotesController {
 	
 	@PostMapping("/save")
 	public String saveNote(@ModelAttribute("notes") Note theNote) {
-
 		notesService.save(theNote);
 
 		return "redirect:/notes/list";
@@ -65,8 +54,7 @@ public class NotesController {
 	
 	
 	@GetMapping("/delete")
-	public String delete(@RequestParam("noteId") int id) {
-
+	public String delete(@RequestParam("noteId") Long id) {
 		notesService.deleteById(id);
 
 		return "redirect:/notes/list";
@@ -75,12 +63,9 @@ public class NotesController {
 	@GetMapping("/search")
 	public String search(@RequestParam("noteName") String name,
 						 Model model) {
-
 		List<Note> notes = notesService.searchBy(name);
-
 		model.addAttribute("notes", notes);
 
 		return "/notes/list-notes";
-		
 	}
 }
