+++ src/main/java/io/github/miwlodar/dao/NotesRepository.java
@@ -5,13 +5,13 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import io.github.miwlodar.entity.Note;
 
-public interface NotesRepository extends JpaRepository<Note, Integer> {
+public interface NotesRepository extends JpaRepository<Note, Long> {
 
 	List<Note> findAllByOrderByTitleAsc();
 	
 	// search by title or content of the note
 	List<Note> findByTitleContainsOrContentContainsAllIgnoreCase(String title, String content);
 
-	// search by id and owner of the note
-	List<Note> findByIdContainsAllIgnoreCase(int id);
+	// search by id of the note
+	List<Note> findByIdContainsAllIgnoreCase(Long id);
 }
